+++ lib/src/screens/home.dart
@@ -8,6 +8,7 @@ import 'package:qrio/src/utils.dart';
 import 'package:qrio/src/widgets/default_popup_menu.dart';
 import 'package:qrio/src/widgets/history.dart';
 import 'package:qrio/src/widgets/qrio.dart';
+import 'package:qrio/src/widgets/unfocus.dart';
 
 class Home extends ConsumerWidget {
   const Home({super.key});
@@ -30,94 +31,98 @@ class Home extends ConsumerWidget {
     const topContentHeight = appBarHeight + qrCodeViewHeight;
     final qrioState = ref.watch(qrioStateProvider);
 
-    return Scaffold(
-      resizeToAvoidBottomInset: false,
-      body: Stack(
-        alignment: AlignmentDirectional.topCenter,
-        children: [
-          SizedBox(
-            height: MediaQuery.of(context).padding.top,
-            child: AppBar(
-              automaticallyImplyLeading: false,
-              backgroundColor: Colors.transparent,
-              surfaceTintColor: Colors.transparent,
-            ),
-          ),
-          const Qrio(),
-          Transform.translate(
-            offset: Offset(0, MediaQuery.of(context).padding.top),
-            child: SizedBox(
-              height: 56,
-              child: Row(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  SvgPicture.asset(
-                    Theme.of(context).brightness == Brightness.dark ||
-                            qrioState.selectedTabIndex == 0
-                        ? 'assets/svg/icon_dark.svg'
-                        : 'assets/svg/icon_light.svg',
-                    width: 24,
-                  ),
-                  const SizedBox(width: 8),
-                  Text(
-                    'QR I/O',
-                    style: TextStyle(
-                      fontSize: 22,
-                      color: Theme.of(context).brightness == Brightness.dark ||
-                              qrioState.selectedTabIndex == 0
-                          ? white
-                          : black,
-                    ),
-                  ),
-                ],
+    return Unfocus(
+      child: Scaffold(
+        resizeToAvoidBottomInset: false,
+        body: Stack(
+          alignment: AlignmentDirectional.topCenter,
+          children: [
+            SizedBox(
+              height: MediaQuery.of(context).padding.top,
+              child: AppBar(
+                automaticallyImplyLeading: false,
+                backgroundColor: Colors.transparent,
+                surfaceTintColor: Colors.transparent,
               ),
             ),
-          ),
-          Transform.translate(
-            offset: Offset(screenWidth / 2 - 28, topPadding),
-            child: const SizedBox(
-              height: 56,
-              width: 56,
-              child: DefaultPopupMenu(),
-            ),
-          ),
-          SizedBox.expand(
-            child: DraggableScrollableSheet(
-              controller: draggableScrollableController,
-              initialChildSize: defaultSheetHeight(context),
-              minChildSize: defaultSheetHeight(context),
-              maxChildSize: 1 - (topPadding + topContentHeight) / screenHeight,
-              snap: true,
-              builder: (context, scrollController) {
-                return Container(
-                  margin: const EdgeInsets.fromLTRB(8, 0, 8, 0),
-                  decoration: BoxDecoration(
-                    color: alphaBlend(
-                      Theme.of(context).colorScheme.primary.withOpacity(0.08),
-                      Theme.of(context).colorScheme.background,
+            const Qrio(),
+            Transform.translate(
+              offset: Offset(0, MediaQuery.of(context).padding.top),
+              child: SizedBox(
+                height: 56,
+                child: Row(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    SvgPicture.asset(
+                      Theme.of(context).brightness == Brightness.dark ||
+                              qrioState.selectedTabIndex == 0
+                          ? 'assets/svg/icon_dark.svg'
+                          : 'assets/svg/icon_light.svg',
+                      width: 24,
                     ),
-                    borderRadius: const BorderRadius.only(
-                      topLeft: Radius.circular(20),
-                      topRight: Radius.circular(20),
+                    const SizedBox(width: 8),
+                    Text(
+                      'QR I/O',
+                      style: TextStyle(
+                        fontSize: 22,
+                        color:
+                            Theme.of(context).brightness == Brightness.dark ||
+                                    qrioState.selectedTabIndex == 0
+                                ? white
+                                : black,
+                      ),
                     ),
-                  ),
-                  child: SingleChildScrollView(
-                    controller: scrollController,
-                    child: const History(),
-                  ),
-                );
-              },
+                  ],
+                ),
+              ),
             ),
-          ),
-          Container(
-            height: bottomPadding,
-            width: MediaQuery.of(context).size.width,
-            margin: EdgeInsets.only(
-              top: MediaQuery.of(context).size.height - bottomPadding,
+            Transform.translate(
+              offset: Offset(screenWidth / 2 - 28, topPadding),
+              child: const SizedBox(
+                height: 56,
+                width: 56,
+                child: DefaultPopupMenu(),
+              ),
             ),
-            color: Colors.transparent,
-          ),
-        ],
+            SizedBox.expand(
+              child: DraggableScrollableSheet(
+                controller: draggableScrollableController,
+                initialChildSize: defaultSheetHeight(context),
+                minChildSize: defaultSheetHeight(context),
+                maxChildSize:
+                    1 - (topPadding + topContentHeight) / screenHeight,
+                snap: true,
+                builder: (context, scrollController) {
+                  return Container(
+                    margin: const EdgeInsets.fromLTRB(8, 0, 8, 0),
+                    decoration: BoxDecoration(
+                      color: alphaBlend(
+                        Theme.of(context).colorScheme.primary.withOpacity(0.08),
+                        Theme.of(context).colorScheme.background,
+                      ),
+                      borderRadius: const BorderRadius.only(
+                        topLeft: Radius.circular(20),
+                        topRight: Radius.circular(20),
+                      ),
+                    ),
+                    child: SingleChildScrollView(
+                      controller: scrollController,
+                      child: const History(),
+                    ),
+                  );
+                },
+              ),
+            ),
+            Container(
+              height: bottomPadding,
+              width: MediaQuery.of(context).size.width,
+              margin: EdgeInsets.only(
+                top: MediaQuery.of(context).size.height - bottomPadding,
+              ),
+              color: Colors.transparent,
+            ),
+          ],
+        ),
       ),
     );
   }
